+++ js/const.js
@@ -1,0 +1,3 @@
+export const GENERATE_OBJECTS = 25;
+export const LIKES = {MIN: 5, MAX:200};
+export const COMMENTS = {MIN: 0, MAX:200};
