+++ js/data.js
@@ -1,0 +1,14 @@
+const descriptionsArr = [
+  'Et ea anim aliqua excepteur ea ut incididunt cupidatat deserunt cillum eu ex ipsum labore.',
+  'Aliqua minim cupidatat veniam id reprehenderit consequat aliquip reprehenderit tempor sunt.',
+  'Anim eiusmod laborum culpa incididunt nisi id ex laboris irure consectetur.',
+  'Eu ea occaecat nulla aute deserunt nulla consectetur.',
+  'Ad anim laborum reprehenderit non irure non labore commodo aute veniam aliquip consectetur.',
+  'Eu Lorem in quis minim officia consequat nisi sint voluptate ex.',
+  'Excepteur occaecat adipisicing aliqua Lorem irure consequat commodo amet adipisicing quis veniam.',
+  'Labore amet excepteur tempor velit amet ipsum qui et anim occaecat labore tempor proident.',
+  'Nostrud anim dolor veniam duis.',
+  'Fugiat adipisicing reprehenderit commodo adipisicing duis.'
+];
+
+export {descriptionsArr};
